etl-pipeline-portfolio/requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.26.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# HTTP requests
requests==2.31.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Data quality
great-expectations==0.18.8

# File processing
chardet==5.2.0
openpyxl==3.1.2

# Airflow (for local development)
apache-airflow==2.8.1

# Logging
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==23.12.1
flake8==7.0.0
pylint==3.0.3
mypy==1.7.1

# Utilities
tqdm==4.66.1
click==8.1.7

# Date handling
python-dateutil==2.8.2

# Development
ipython==8.19.0
jupyter==1.0.0
notebook==7.0.6
